Fixed Repeated Loading of Animation
The code had an issue where every time the cat changed sides the animation was being loaded. These redundant loads probably don't effect performance much in a game this small, but it is still better to create the reference to the loaded animation once and then run it each time the cat changes sides.